Cleanup top-level Makefile.
This should allow the Makefile to be used to create a working
Sancus development environment on a clean Ubuntu 16.04 machine.

Note: LLVM/Clang can either be installed via a pre-built Debian
package, or built from source.

Note: got rid of sancus-* git submodules, such that we can always test the
latest versions that are explicitly pulled.